Moving SteerDirection logic to Steering Behaviour Class Creating Vector2Extensions for rotate vector
