+++ Runtime/Common/Utils/Extensions/Vector2Extensions.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 059aeb69bee3e6c4682f45fc5155e298
+guid: 6d17f5359da634d42bcd8ec160f004d4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
